+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * Add ability to search by barcode. Part of UIIN-143.
 * Move acquisitions items to the acquisitions accordion. Fixes UIIN-506.
 * Remove duplicate asterisks from required fields. Fixes UIIN-490.
+* Add ability to mark Paged items as missing. Part of UIIN-555.
 
 ## [1.7.0](https://github.com/folio-org/ui-inventory/tree/v1.7.0) (2019-03-15)
 [Full Changelog](https://github.com/folio-org/ui-inventory/compare/v1.6.0...v1.7.0)